requirements.txt
keboola.component==1.4.3
snowflake-connector-python
git+https://github.com/bakobako/dataconf#egg=dataconf
mock
freezegun

